update users, posrs, base.html

+++ CH3_Assignment/posts/views.py
@@ -1,6 +1,11 @@
 from django.shortcuts import get_object_or_404, render, redirect
+from django.contrib.auth.decorators import login_required
+from django.views.decorators.http import require_POST, require_http_methods
 from .forms import PostForm
 from .models import Post
+
+def home(request) :
+    return render(request, 'posts/home.html')
 
 def post_list(request) :
     posts = Post.objects.all().order_by('-pk')
@@ -18,7 +23,7 @@ def post_detail(requset, pk) :
     return render(requset, 'posts/post_detail.html', context)
 
 
-
+@login_required
 def post_create(request):
     if request.method == 'POST' :
         form = PostForm(request.POST)
@@ -26,7 +31,7 @@ def post_create(request):
             post = form.save()
             return redirect('posts:post_detail', post.pk)
     else :
-        form = PostForm
+        form = PostForm()
 
     context = {
         'form' : form,
@@ -35,14 +40,14 @@ def post_create(request):
 
 def post_update(request, pk):
     post = get_object_or_404(Post, pk=pk)
-    print(post.title)
-    if request.method == 'POST':
-        form = PostForm(request.POST, instance=post)
-        if form.is_valid():
-            post = form.save()
-            return redirect('posts:post_detail', pk=post.pk)
-    else:
-        form = PostForm(instance=post)
+    if post.author== request.user:
+        if request.method == 'POST':
+            form = PostForm(request.POST, instance=post)
+            if form.is_valid():
+                post = form.save()
+                return redirect('posts:post_detail', pk=post.pk)
+        else:
+            form = PostForm(instance=post)
     
     context = {
         'form': form,
@@ -50,18 +55,21 @@ def post_update(request, pk):
     }
     return render(request, 'posts/post_form.html', context)
 
-
-def post_delete(request, pk) : 
-    # 삭제할 post 객체를 가져오기
+@require_http_methods(["GET", "POST"])
+def post_delete(request, pk):
     post = get_object_or_404(Post, pk=pk)
+    
+    # 요청자가 작성자인지 확인
+    if post.author != request.user:
+        return redirect("posts:post_list")  # 권한이 없으면 목록 페이지로 리디렉션
 
-    # POST 요청이 오면 삭제
     if request.method == "POST":
+        # POST 요청 시 삭제 수행
         post.delete()
-        # 삭제 후에 목록 페이지로
         return redirect("posts:post_list")
-    # GET 요청이 오면 삭제 확인 페이지
+    
+    # GET 요청 시 삭제 확인 페이지 렌더링
     context = {
-        'post' : post
+        'post': post
     }
     return render(request, "posts/post_confirm_delete.html", context)
